requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.2
attrs==20.3.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
boto3==1.14.9
botocore==1.17.9
bs4==0.0.1
cached-property==1.5.2
cachetools==4.1.1
celery==4.4.7
certifi==2020.4.5.2
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
configparser==5.0.1
crayons==0.4.0
cryptography==3.0
cssselect==1.1.0
dateparser==1.0.0
defusedxml==0.7.0rc1
distlib==0.3.0
Django==3.0.5
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.4.0
django-csp==3.6
django-dynamic-db-router==0.3.0
django-filter==2.3.0
django-ipware==3.0.1
django-magiclink==0.8.0
django-mercadopago==5.1.2
django-money==1.1
django-social-share==2.0.0
django-storages==1.9.1
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djongo==1.3.3
dnspython==1.16.0
docutils==0.15.2
environ==1.0
facebook-scraper==0.2.17
fake-useragent==0.1.11
filelock==3.0.12
google-auth==1.20.1
google-auth-oauthlib==0.4.1
googlemaps==4.4.2
gspread==3.6.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
importlib-metadata==1.6.1
isodate==0.6.0
isort==4.3.21
jmespath==0.10.0
kombu==4.6.11
lazy-object-proxy==1.4.3
lxml==4.6.1
Markdown==3.2.2
mccabe==0.6.1
mercadopago==1.1.1
oauth2client==4.1.3
oauthlib==3.1.0
parse==1.18.0
PGPy==0.5.3
Pillow==7.2.0
pipenv==2020.6.2
postgis==1.0.4
psycopg2==2.8.5
psycopg2-binary==2.8.5
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyee==7.0.4
PyJWT==1.7.1
pylint==2.5.3
pymongo==3.11.0
pyppeteer==0.2.2
pyquery==1.4.3
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.11.13
reportlab==3.5.46
requests==2.23.0
requests-file==1.5.1
requests-html==0.10.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
s3transfer==0.3.3
selenium==3.141.0
setuptools-git==1.2
setuptools-scm==4.1.2
signals==0.0.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.2.4
toml==0.10.1
tqdm==4.53.0
typed-ast==1.4.1
tzlocal==2.1
-e git+https://github.com/harismuneer/Ultimate-Facebook-Scraper.git@4f54d8473d35b16f81ba5fbff7f112acc35c57bc#egg=ultimate_facebook_scraper
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.21
virtualenv-clone==0.5.4
w3lib==1.22.0
webdriver-manager==3.2.2
websockets==8.1
wrapt==1.12.1
zeep==4.0.0
zipp==3.1.0
